+++ components/organism/AdminTabs/AllLinks.jsx
@@ -1,11 +1,11 @@
-import Pagination from '../../pagination';
+import Pagination from "../Pagination";
 
 const AllLinks = () => {
-    return (
-        <>
-            <Pagination slug="allLinks" />
-        </>
-    );
+  return (
+    <>
+      <Pagination slug="allLinks" />
+    </>
+  );
 };
 
 export default AllLinks;
