+++ components/organism/AdminTabs/MyLinks.jsx
@@ -1,11 +1,11 @@
-import Pagination from '../../pagination';
+import Pagination from "../Pagination";
 
 const MyLinks = () => {
-    return (
-        <>
-            <Pagination slug="myLinks" />
-        </>
-    );
+  return (
+    <>
+      <Pagination slug="myLinks" />
+    </>
+  );
 };
 
 export default MyLinks;
